+++ addOns/client/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Wait always after component navigation while crawling.
 - Follow any navigation component while crawling.
+- Crawl component state changes.
 
 ### Fixed
 - Normalise behaviour of Delete context menu item.